+++ __tests__/__utils__/handlers.ts
@@ -24,7 +24,6 @@ import * as R from 'ramda'
 
 import { Model } from '~/internals/graphql'
 import { Payload } from '~/internals/model/types'
-import { NotificationEventPayload } from '~/schema/notification/types'
 import { NavigationPayload } from '~/schema/uuid/abstract-navigation-child/types'
 import { UuidPayload } from '~/schema/uuid/abstract-uuid/types'
 
@@ -64,7 +63,7 @@ export function createNavigationHandler(navigation: NavigationPayload) {
 }
 
 export function createNotificationEventHandler(
-  notificationEvent: NotificationEventPayload
+  notificationEvent: Model<'AbstractNotificationEvent'>
 ) {
   return createMessageHandler({
     message: {