+++ src/api.ts
@@ -19,18 +19,15 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-export {
-  AuthorizationGuard,
-  AuthorizationPayload,
-  Permission,
-  Scope,
-} from '~/authorization'
 export * from '~/internals/authentication/service'
 export * from '~/internals/cache'
 export * from '~/internals/data-source'
 export * from '~/internals/environment'
 export * from '~/internals/error-event'
-export * from '~/internals/graphql'
+export * from '~/internals/graphql/context'
+export * from '~/internals/graphql/schema'
+export * from '~/internals/graphql/resolver'
+export { createMutationNamespace } from '~/internals/graphql/utils'
 export * from '~/internals/model/query'
 export * from '~/internals/model/types'
 export * from '~/internals/swr-queue'
@@ -39,23 +36,6 @@ export * from '~/model'
 export * from '~/model/decoder'
 export * from '~/model/types'
 export * from '~/schema/connection/types'
-export * from '~/schema/notification/types'
-export * from '~/schema/notification/checkout-revision-notification-event/types'
-export * from '~/schema/notification/create-comment-notification-event/types'
-export * from '~/schema/notification/create-entity-link-notification-event/types'
-export * from '~/schema/notification/create-entity-notification-event/types'
-export * from '~/schema/notification/create-entity-revision-notification-event/types'
-export * from '~/schema/notification/create-taxonomy-link-notification-event/types'
-export * from '~/schema/notification/create-taxonomy-term-notification-event/types'
-export * from '~/schema/notification/create-thread-notification-event/types'
-export * from '~/schema/notification/reject-revision-notification-event/types'
-export * from '~/schema/notification/remove-entity-link-notification-event/types'
-export * from '~/schema/notification/remove-taxonomy-link-notification-event/types'
-export * from '~/schema/notification/set-license-notification-event/types'
-export * from '~/schema/notification/set-taxonomy-parent-notification-event/types'
-export * from '~/schema/notification/set-taxonomy-term-notification-event/types'
-export * from '~/schema/notification/set-thread-state-notification-event/types'
-export * from '~/schema/notification/set-uuid-state-notification-event/types'
 export * from '~/schema/subscription/types'
 export * from '~/schema/uuid/abstract-navigation-child/types'
 export * from '~/schema/uuid/abstract-repository/types'