+++ src/schema/notification/create-entity-notification-event/resolvers.ts
@@ -20,16 +20,22 @@
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
 import { createNotificationEventResolvers } from '../utils'
-import { LegacyCreateEntityNotificationEventResolvers } from './types'
-import { EntityPayload } from '~/schema/uuid/abstract-entity/types'
+import { TypeResolvers } from '~/internals/graphql'
+import { EntityDecoder } from '~/model/decoder'
+import { CreateEntityNotificationEvent } from '~/types'
 
-export const resolvers: LegacyCreateEntityNotificationEventResolvers = {
+export const resolvers: TypeResolvers<CreateEntityNotificationEvent> = {
   CreateEntityNotificationEvent: {
     ...createNotificationEventResolvers(),
     async entity(notificationEvent, _args, { dataSources }) {
-      return (await dataSources.model.serlo.getUuid({
+      const entity = await dataSources.model.serlo.getUuidWithCustomDecoder({
         id: notificationEvent.entityId,
-      })) as EntityPayload | null
+        decoder: EntityDecoder,
+      })
+
+      if (entity === null) throw new Error('entity cannot be null')
+
+      return entity
     },
   },
 }